backend/requirements.txt
Flask
requests
beautifulsoup4
openai
langchain
faiss-cpu
python-dotenv
flask-cors
tiktoken
google-cloud-logging
google-cloud-storage